SerializableTransactionManagerImpl -> SerializableTransactionManager
