--- atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/transaction/impl/SnapshotTransactionManager.java
@@ -1,43 +1,0 @@
-/*
- * Copyright 2018 Palantir Technologies, Inc. All rights reserved.
- *
- * Licensed under the BSD-3 License (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://opensource.org/licenses/BSD-3-Clause
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.palantir.atlasdb.transaction.impl;
-
-import com.palantir.atlasdb.cleaner.Cleaner;
-import com.palantir.atlasdb.transaction.api.PreCommitCondition;
-import com.palantir.atlasdb.transaction.api.TransactionFailedRetriableException;
-import com.palantir.atlasdb.transaction.api.TransactionTask;
-import com.palantir.timestamp.TimestampService;
-
-public interface SnapshotTransactionManager {
-    RawTransaction setupRunTaskWithConditionThrowOnConflict(PreCommitCondition condition);
-
-    <T, E extends Exception> T finishRunTaskWithLockThrowOnConflict(RawTransaction tx,
-            TransactionTask<T, E> task)
-            throws E, TransactionFailedRetriableException;
-
-    /**
-     * Registers a Runnable that will be run when the transaction manager is closed, provided no callback already
-     * submitted throws an exception.
-     *
-     * Concurrency: If this method races with close(), then closingCallback may not be called.
-     */
-    void registerClosingCallback(Runnable closingCallback);
-
-    Cleaner getCleaner();
-
-    TimestampService getTimestampService();
-}
